+++ frontend/package.json
@@ -13,10 +13,12 @@
   "dependencies": {
     "axios": "^1.8.4",
     "lucide-react": "^0.483.0",
+    "react-toastify": "^11.0.5",
+    "helmet": "^8.1.0",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
-    "react-router-dom": "^7.4.0",
-    "react-toastify": "^11.0.5"
+    "react-helmet": "^6.1.0",
+    "react-router-dom": "^7.4.0"
   },
   "devDependencies": {
     "@eslint/js": "^9.19.0",
