+++ frontend/src/page/administrador/VentanaEmergente.css
@@ -1,85 +1,85 @@
 /* VentanaEmergente.css */
 .ventana-emergente {
-    position: fixed;
-    top: 0;
-    left: 0;
-    right: 0;
-    bottom: 0;
-    background-color: rgba(0, 0, 0, 0.5); /* Fondo semi-transparente */
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    z-index: 1000; /* Asegúrate de que esté por encima de otros elementos */
-  }
-  
-  .contenido-ventana {
-    background-color: white;
-    padding: 20px;
-    border-radius: 5px;
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2);
-    text-align: center;
-    width: 600px;
-height: 97px;
-    position: relative; /* Para posicionar el botón "X" en relación con el contenedor */
-    border: 3px solid rgb(15, 2, 66); /* Borde azul oscuro */
-  }
-  
-  .btn-cerrar {
-    position: absolute;
-    top: 10px; /* Espaciado desde la parte superior */
-    right: 10px; /* Espaciado desde la derecha */
-    background-color: red; /* Color de fondo rojo */
-    color: white; /* Color del texto */
-    border: none; /* Sin borde */
-    border-radius: 50%; /* Botón redondeado */
-    width: 25px; /* Ancho del botón */
-    height: 25px; /* Alto del botón */
-    cursor: pointer; /* Cambia el cursor al pasar el mouse */
-    font-size: 16px; /* Tamaño del texto */
-  }
-  
-  .btn-cerrar:hover {
-    background-color: darkred; /* Color al pasar el mouse */
-  }
-  
-  .botones {
-    margin-top: 20px;
-    display: flex;
-    justify-content: center; /* Centra los botones */
-    gap: 10px; /* Espacio entre los botones */
-  }
-  
-  .btn-confirmar {
-    width: 176px; /* Ancho del botón */
-    height: 37px; /* Alto del botón */
-    display: flex;
-    align-items: center;
-    justify-content: center; /* Centra el texto dentro del botón */
-    border-radius: 15px;
-    background-color: rgba(55, 79, 214, 1); /* Color de fondo solo para el botón "Sí" */
-    box-shadow: 0px 4px 4px 0px rgba(0, 0, 0, 0.25); /* Sombra */
-    border: none; /* Sin borde */
-    cursor: pointer; /* Cambia el cursor al pasar el mouse */
-  }
-  
-  .btn-cancelar {
-    width: 176px; /* Ancho del botón */
-    height: 37px; /* Alto del botón */
-    display: flex;
-    align-items: center;
-    justify-content: center; /* Centra el texto dentro del botón */
-    border-radius: 15px;
-    background-color: rgba(222, 225, 230, 1); /* Color de fondo solo para el botón "No" */
-    box-shadow: 0px 4px 4px 0px rgba(0, 0, 0, 0.25); /* Sombra */
-    border: none; /* Sin borde */
-    cursor: pointer; /* Cambia el cursor al pasar el mouse */
-  }
-  
-  .btn-confirmar:hover {
-    background-color: rgba(45, 60, 175, 1); /* Color al pasar el mouse para "Sí" */
-  }
-  
-  .btn-cancelar:hover {
-    background-color: rgba(200, 205, 210, 1); /* Color al pasar el mouse para "No" */
-  }
-  
+  position: fixed;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  background-color: rgba(0, 0, 0, 0.5); /* Fondo semi-transparente */
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  z-index: 1000; /* Asegúrate de que esté por encima de otros elementos */
+}
+
+.contenido-ventana {
+  background-color: white;
+  padding: 20px;
+  border-radius: 5px;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2);
+  text-align: center;
+  width: 600px;
+height: 125px;
+  position: relative; /* Para posicionar el botón "X" en relación con el contenedor */
+  border: 3px solid rgb(15, 2, 66); /* Borde azul oscuro */
+}
+
+.btn-cerrar {
+  position: absolute;
+  top: 10px; /* Espaciado desde la parte superior */
+  right: 10px; /* Espaciado desde la derecha */
+  background-color: red; /* Color de fondo rojo */
+  color: white; /* Color del texto */
+  border: none; /* Sin borde */
+  border-radius: 50%; /* Botón redondeado */
+  width: 25px; /* Ancho del botón */
+  height: 25px; /* Alto del botón */
+  cursor: pointer; /* Cambia el cursor al pasar el mouse */
+  font-size: 16px; /* Tamaño del texto */
+}
+
+.btn-cerrar:hover {
+  background-color: darkred; /* Color al pasar el mouse */
+}
+
+.botones {
+  margin-top: 20px;
+  display: flex;
+  justify-content: center; /* Centra los botones */
+  gap: 10px; /* Espacio entre los botones */
+}
+
+.btn-confirmar {
+  width: 176px; /* Ancho del botón */
+  height: 37px; /* Alto del botón */
+  display: flex;
+  align-items: center;
+  justify-content: center; /* Centra el texto dentro del botón */
+  border-radius: 15px;
+  color: white;
+  background-color: #132a9c; /* Color de fondo solo para el botón "Sí" */
+  box-shadow: 0px 4px 4px 0px rgba(0, 0, 0, 0.25); /* Sombra */
+  border: none; /* Sin borde */
+  cursor: pointer; /* Cambia el cursor al pasar el mouse */
+}
+
+.btn-cancelar {
+  width: 176px; /* Ancho del botón */
+  height: 37px; /* Alto del botón */
+  display: flex;
+  align-items: center;
+  justify-content: center; /* Centra el texto dentro del botón */
+  border-radius: 15px;
+  background-color: rgba(222, 225, 230, 1); /* Color de fondo solo para el botón "No" */
+  box-shadow: 0px 4px 4px 0px rgba(0, 0, 0, 0.25); /* Sombra */
+  border: none; /* Sin borde */
+  cursor: pointer; /* Cambia el cursor al pasar el mouse */
+}
+
+.btn-confirmar:hover {
+  background-color: rgba(45, 60, 175, 1); /* Color al pasar el mouse para "Sí" */
+}
+
+.btn-cancelar:hover {
+  background-color: rgba(200, 205, 210, 1); /* Color al pasar el mouse para "No" */
+}
